Update README.md
Setting the networking part on MacOS did not work with the setting on Linux ubuntu. The best I could find is to add a second adapter and set it to Host-only Adapter. When a VM is started an IP address is assigned to it 192.168.1.XXX. Log using tester account and get the acctual address using ifconfig